pyproject.toml
[project]
name = "mcp-gilbut"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.1.0",
    "fastapi>=0.116.2",
    "fastmcp>=2.12.3",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.33",
    "mcp>=1.14.0",
    "openai>=1.107.3",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pypdf>=6.0.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "tiktoken>=0.11.0",
    "uvicorn>=0.35.0",
]
